Contexts: fix orphaned browser windows failing tests.
After scenario, we instruct Mink to reset the browser environment.
It will try to clear the cookies and reset the request headers and the
browser history to the limit of the Mink webdriver possibilities.

Instead of a reset, this change forces Mink to close and re-open the browser.

Behaviour has been seen where a Selenium tests leaves open a browser window
after a test is complete, and not close it. The 'after scenario' will clear
session cookies in that browser, and eventually, the WordPress application's
heartbeat API will render a blocking modal showing "you must log-in again".

This can interfere with subsequent tests in that browser, as visiting a URL
can sometimes cause a browser "are you sure you want to leave this page" prompt
to appear which breaks navigation.

It also confuses the user running about if the tests all actually passed.